Extract getApplicantDetails to common applicationUtils
